19.2.v-if, v-show and Conditional Classes{ inspecting v-if and the v-show directive }
inspecting v-if and the v-show directive
The other way to hide things is by using v-show. With v-show things will look the same. The difference is that, when we click collapse, the HTML is still there. v-show just adds display: none to the element when it needs to. Sneaky!
So v-show and v-if accomplish the same thing. Which you should use just depends on the situation. v-show is usually better if you're hiding and showing a lot, like a drop-down menu. That's because v-show is super fast.
But if what you're hiding isn't shown often or is complex - like a modal - v-if might be better because Vue won't spend any time even rendering the element until you need it.