19.3.v-if, v-show and Conditional Classes{ .component style }
.component style
Ok! While we're here, I also want to clean things up. Let's refactor our width away from style to a proper class. But before we do that, there's one thing I wanted to do earlier that I forgot about. Notice that down here in our styles, we have a single class called sidebar... and we're applying that to the root element of our template. This is great! But sometimes, as a convention, when you have a class that is applied to the root element of a component, you'll use the generic name .component.
Down in style, update the class name to match.
This won't make any difference, it's just a nice standard when you need a class on your outer element.
To remove this width style, go back down to the style block and, inside .component add a new class called &.collapsed with width: 70px.
If you're not familiar with this syntax, it means that if an element has both the component and the collapsed classes, then it will get this width.
Back in the template, delete the style attribute.
The tricky part here is that we need to conditionally add the collapsed class to our root element. But... there's not really a good way to do that.