19.4.v-if, v-show and Conditional Classes{ Conditional :class }
Conditional :class
Luckily Vue is here again to rescue us! It turns out, :class has another superhero syntax.
Instead of an array, pass an object where every key in the object will be the class name that you want to add, and the value will be true or false for whether or not you want this class. Since we always want these three classes, we'll say $style.component: true, 'p-3': true and 'mb-5': true.
And... Webpack is furious with me! The reason is that, in an object, you can't use variable names as keys. The problem is that JavaScript thinks that all keys are strings, even if you don't have quotes around them.