19.5.v-if, v-show and Conditional Classes{ Conditional :class }
 If you want tell JavaScript that your key is some sort of variable, you need to put square brackets around it. That's a bit ugly, but now JavaScript understands it.
To add the collapsed class conditionally. I'll use that funny syntax again to say $style.collapsed, because every time we reference a class within modular CSS, we need to use $style.. And since we want the collapsed class to show if the collapsed state is true, we'll just set it to collapsed.
Back in the browser, click on >> Collapse and... perfect! You can see our class hiding and showing in the DOM.
But... there's a slightly better way to accomplish all of this... and it will use one of my favorite features of Vue: computed properties. Let's talk about them next.