20.2.Computed Properties{ How computed works; Check it out! }
Anyways, because computed keys are added to the Vue instance, just like methods, data or props, it means that we can reference them with the this variable. To prove it, inside of our toggleCollapsed method, let's say console.log(this.componentClass).
Notice that PhpStorm tries to autocomplete that with parentheses, but that's not right! We need to reference it like a property, even though we know that Vue will call our method.

Check it out!
So if we go over to the browser now... check this out! You can see the log and it shows the component classes correctly every time we change it!
You can also see this over in the Vue dev tools! Click down on Sidebar. Under data, you now have a computed section with componentClasses. This changes when you hide and show the sidebar!