20.3.Computed Properties{ Check it out! }
Back at our editor, remove the console.log(). We just mastered one of the most powerful tools in Vue: computed properties! Nice work!
Next, what happens when we need to access a piece of data - like collapsed - in a different component? If that component is a child, we can pass it down as a prop. But... what if it's not?