22.1 Communication UP with $emit {Communicating to a Parent Component with $emit()}
The sidebarCollapsed data now lives inside of Products because, in a minute, we're going to use it to dynamically change the classes on the sidebar and content elements. We're also already passing this data as a collapsed prop into Sidebar so we can happily reference it inside the template.
The problem is that, when we click the Collapse button, what we really want to do is change the data on the parent component: we want to change sidebarCollapsed on products.vue. But... you can't do that. A component can only change data on itself.
Here's the deal: we already know how to communicate information down the component tree. We do that via props: we communicate the sidebarCollapsed state down into sidebar with a prop. Cool! But how can we communicate up the tree?

Communicating to a Parent Component with ()
Whenever a child component needs to change data on a parent component... or, more abstractly: whenever a child component needs to communicate that something happened to a parent component - like "the collapsed button was clicked" - it does that by emitting an event.
Check this out: inside of sidebar, go down to the toggleCollapsed() method that's being called on click. Instead of modifying the prop, say this.$emit(). Yes, in addition to props, data, methods and computed props, the Vue instance - the this variable - also has a few built-in methods and properties. There aren't a ton of methods, but there is one called $emit and it's one of the most important.
Inside of $emit(), let's emit a custom event called toggle-collapsed, which I totally just made up.
This won't work yet, but we can already see it in action. Back on your browser, over in the Vue dev tools, select the Sidebar component. So far, we've been paying a lot of attention to the props, data and computed info. But there is also a section called events. This will show any events that our Vue components are emitting. So now, when we hit the Collapse button, nothing happens yet, but we can see the toggle-collapsed event being emitted! That's pretty cool. The next step will be to listen to this event from the parent component.