22.2 Communication UP with $emit {Emitting Directly in v-on; Listening to the Event on the Parent Component}
Emitting Directly in v-on
Before we do that, calling this.() is totally fine from inside a method. But we can simplify. Copy the $emit() code then delete the method entirely. Up in the template, find the @click, which we know is really v-on:click. Set it to $emit('toggle-collapsed').
Because... remember! Whenever you reference a variable or function inside a template, Vue will really call this.() behind the scenes... which is exactly what we were doing a moment ago. This is just shorter... and it emits the event just like before.

Listening to the Event on the Parent Component
The second step to this process is to go into the products component and listen to that event. I'll move the <sidebar onto multiple lines.
To listen to the toggle-collapsed event, we're going to use v-on. Because, really, listening to a custom event is no different than what we're doing in sidebar. To listen to the click event of a button, we use v-on:click, or @click for short. Then, on click, we run some code!
click is a native DOM event, but things work exactly the same for an event that we're emitting manually.
Let's do this the long way first: say v-on:toggle-collapsed= and set this to call a new toggleSidebarCollapsed method when that event happens.
Copy that name and go down to the script section. We don't have any methods yet so first add methods, then create the toggleSidebarCollapsed() function inside. Very simply - just like we did before in sidebar - set this.sidebarCollapsed to ! this.sidebarCollapsed.
I love that. Back on the browser, I'll refresh to be safe... and then click Collapse. It works! You can see the events and, back on the Products component, we can watch the sidebarCollapsed data change. That is a nice setup.