22.3 Communication UP with $emit {Using Shortcut v-on Everywhere}
Using Shortcut v-on Everywhere
Now that we have this working, go back and find the v-on: attribute. What I love about this directive is how clear it is:

  on toggle-collapsed, run this code

But in practice, we're going to use the shortcut syntax everywhere. So: @toggle-collapsed.
Just remember that the @ symbol means "on" - so "on toggle collapsed".
Our communication between the Products component and Sidebar - using props to communicate down and events to communicate up - is a pattern that you'll see many times in Vue. It keeps a single source of data, but effectively allows anyone to read or change it.
Next, let's use the new sidebarCollapsed data in products.vue. When we do that, we're going to add some custom styles but decide to not make them modular.