24.2. Ajax with Axios{Making the AJAX Call from mounted()}
Making the AJAX Call from mounted()
Ok, when we load products, which component is going to need that data? The top-level products.vue component renders the sidebar and catalog. We could load the products here... but we won't actually need them. Hold Command or Ctrl and click <catalog to jump to that component.
Ah, this is the component that needs the products data.
Here's the goal: as soon as Vue loads this component, we'll start the AJAX call so that we can load the products as quickly as possible. Fortunately, Vue allows us to run code during its startup process, and there are two main "hook" points: mounted and created. We'll talk more about these later but Vue considers your component mounted when it's actually added to the page - like, in products.js when we call .$mount().
To run code right after our component is mounted, all we need to do is create a function called mounted(). Inside, we'll make the AJAX call.
How? First, at the top of the script section import axios from 'axios'.
Then, using Axios is beautifully simple: axios.get('/api/products'). And like every AJAX library, this will return a Promise, which you can learn all about in a JavaScript Tutorial here on SymfonyCasts.
To use the Promise, add .then(), and pass an arrow function with response as the argument. Let's console.log(response) to see what it looks like.
Testing time! Back over on the browser, click to view the console. Thanks to hot module replacement... that already ran! But to make the flow more realistic, let's refresh the page.
Now... boom! The log shows up almost instantly. The response is an object with headers, status and other things. What we want is data. One of the nice features of Axios is that it decodes the JSON automatically.
When you're working with JSON-LD Hydra like this, the collection of items is stored on a hydra:member property. Yep, it's an array with 12 products. We have product data!
Next, this is working great, but I'm going to choose a slightly different syntax for handling Promises: async and await. Then, we'll use our brand new data to render those products onto the page.