25.1. The await Keyword
Before we start using the products data from the AJAX call, there's one other way to work with promises... and I really like it! It's the await syntax. We know that Axios returns a Promise... and that we normally run code after a promise has finished - or resolved - by calling .then() on it.
This works great. But instead add const response = before the axios call and then remove the callback.
If we stopped right now, response would actually be a Promise - not a response. But if we put await in front of it, it will be a response! The await keyword causes your code to wait for that Promise to resolve. And whatever data is normally passed to your callback as an argument is instead returned. There is still an asynchronous AJAX call happening, but our code reads a bit more like synchronous code. The await keyword is syntactic sugar.
Hmm... PhpStorm looks mad... but... let's ignore it and try this anyways! Move back over to the browser and scroll to the bottom of the console. Ah!

  Cannot use keyword await outside an async function.

So... first: I hope that the general idea of await makes sense to you. If you have something that is asynchronous, you put await in front of it and that says:

  Please wait for this to finish, get the return value and then keep executing my code.