25.2. The await Keyword{Why do we need async?}
Why do we need async?

That's great. But this comes with one rule: whenever you use the await keyword, whatever function you're inside of needs to have an async keyword in front of it. Let me put back the console.log(response).
When you make a function async, it means that your function will now automatically and always return a Promise. If your function has a return value, that will be the data of the Promise.
This... can be confusing at first: when mounted() is called, our code will freeze on the async line and wait for the AJAX call to finish. But this doesn't freeze our entire JavaScript app. In reality, the mounted() function will almost immediately finish and will return a Promise. That Promise will resolve once all of our code executes.
To say this a different way: if we called mounted() directly from our code - we won't do that, but just pretend - then mounted() would finish before the AJAX call and it would now return a Promise. If we wanted to do something after the AJAX call and the rest of the code in mounted() finished, we could chain a .then() from that Promise.
But in reality, Vue is responsible for calling mounted() and Vue doesn't care about or use any value that we might return from mounted(). So basically, Vue couldn't care less that we just changed this to async and so, caused our method to return a Promise.
The key thing to know about async and await is that even though our code will wait on this line, really the mounted() function will finish nearly instantly. Vue isn't going to call mounted() then freeze our entire Vue app waiting for it to finish. It starts our AJAX call then keeps going. That's perfect.
Anyways, now when we refresh... yes! The AJAX call finishes and logs the response. Feel free to use Promises directly or with await: we'll use await in this tutorial.