25.3. The await Keyword{ Adding the products Data & hydra:member; The Useful @id IRI }
Adding the products Data & hydra:member
Ok, let's use our real products data! Inside catalog, if you think about it, the products are something that will change during the lifecycle of our component. At the very least, when the catalog component first loads, the products will be empty. And then, as soon as the Ajax call finishes, they will change to be the real products.
That's a long way of saying that products need live in data. Head up to to the data option and add a new products key set to an empty array as its default data.
In mounted(), instead of logging the response, now we can say this.products = response.data and... let's see. Inside response.data, we want the hydra:member property. Cool! Add .hydra:member. Oh but that : mucks things up. We need to use square brackets here with that in quotes
Ok, let's check out the products data. Move over... I'll refresh just to be safe... then on the Vue dev tools, find the Catalog component. Yes! Our products data has 12 items in it!
So... let's celebrate and use this data! Up in the template, remove that pesky TODO. In our design, we need to have one of these divs for each product. Break it into multiple lines and then loop with v-for="product in products". And every time we use v-for, we need to add a key attribute set to some unique, non-changing key for each item.

The Useful @id IRI
If you look at the products data on the dev tools, you'll notice that we do have an id property. It's not very important in this case, but I'm actually going to use this @id instead. This is called an "IRI" - it's a unique key that API Platform adds to every resource. It's... just more useful than a database ID or UUID because it's a real URL: we could make a request to this address to fetch that specific product.
That "usefulness" won't be... well... useful in this situation. But because of this, in general, I'm going to use @id everywhere as my unique identifier. The only problem is that when you say product.@id, JavaScript gets mad because you can't use an @ sign in this syntax. Once again, use square brackets and wrap this in quotes.
Inside the div, start by printing the name: {{ product.name }}... because name is one of the keys in the data. And... yea! You can already see it being used! There is our list of real, high-quality, products.
Next, if we added all of the markup and data we need for each product directly into the catalog component... things could get big and ugly fast. Let's split the product listing into two, smaller, sleeker components.