26.2. Product Listing Components{ Creating a product-card Component }
Creating a product-card Component
And once again, we could stop now and start adding more product info right here. But I'm going to go one step further and refactor each individual product into its own component.
Inside the product-list/ directory, create a second component called, how about, product-card.vue. Start like we always do: with the <template>. Each product will use use this div with the col classes on it. Copy the classes but keep the v-for here. In product-card, add a div and paste those classes.
Let's also go steal the {{ product.name }} and put that here.
That's a good enough start. Next add the <script> tag with export default and name set to ProductCard. And once again, we know that we're going to need a product passed to us. So I'll jump straight to saying props with product inside. The product will be an object... and like with the Array, there's not an easy way to enforce exactly what that object looks like. But we can at least say type: Object and also required: true in case I do something silly and forget to pass the prop entirely.
Ok! This is ready! Back in index.vue, we'll follow the same process to use the new component. This starts to feel a bit repetitive.. and I love that! It means we're getting comfortable.
Start with import ProductCard from and you can either say ./product-card, which is shorter, or @/components/product-list/product-card, which is more portable.
To make this component available in the template, add the components key with ProductCard inside.
Finally, in the template, this is pretty cool: we now want to loop over the products and render the ProductCard itself each time. And putting a v-for on a custom component is totally legal. Replace the <div> with <product-card>, remove the class attribute and replace it with :product="product". And since this element no longer has any content, it can be self-closing.
That's really nice: we loop over products and render an individual ProductCard for each one. When we check the browser, it looks like it's working! I'll refresh just to be sure... I don't always trust hot module replacement. And... yep! It does work.
What I really like about the product-card component is that it's focused on rendering just one thing: a single product. Next, let's really bring the product to life by adding more data, styles and a computed property.